ENG-337: Widen to a long to fix casting errors.
git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@23757 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ toolkit-impl/src/main/java/com/terracotta/toolkit/config/cache/InternalCacheConfigurationType.java
@@ -618,7 +618,7 @@ public enum InternalCacheConfigurationType {
     return Validator.integer(name(), value);
   }
 
-  public int integerOrLong(Object value) {
+  public long integerOrLong(Object value) {
     return Validator.integerOrLong(name(), value);
   }
